+++ plugins/storage/1701.go
@@ -1,4 +1,4 @@
-package dev
+package storage
 
 import (
 	"strings"
@@ -27,9 +27,9 @@ func init() {
 
 	extractorConfig.Apply(engine)
 
-	engine.OnResponse(func(response *colly.Response, ctx *crawlers.Context) {
-		println(string(response.Body))
-	})
+	//engine.OnResponse(func(response *colly.Response, ctx *crawlers.Context) {
+	//	println(string(response.Body))
+	//})
 
 	engine.OnXML("//loc", func(element *colly.XMLElement, ctx *crawlers.Context) {
 		if strings.HasSuffix(element.Text, ".xml") {
