+++ plugins/storage/1710.go
@@ -1,4 +1,4 @@
-package dev
+package storage
 
 import (
 	"megaCrawler/crawlers"
@@ -8,9 +8,13 @@ import (
 )
 
 func init() {
-	engine := crawlers.Register("1712", "科技部", "https://www.dost.gov.ph/")
+	engine := crawlers.Register("1710", "技术教育和技能发展署", "https://www.tesda.gov.ph/")
 
-	engine.SetStartingURLs([]string{"https://www.dost.gov.ph/"})
+	engine.SetStartingURLs([]string{
+		"https://www.tesda.gov.ph/Media",
+		"https://www.tesda.gov.ph/Media/Events",
+		"https://www.tesda.gov.ph/Media/ProvincialNews",
+	})
 
 	extractorConfig := extractors.Config{
 		Author:       true,
@@ -29,7 +33,7 @@ func init() {
 		engine.Visit(element.Attr("href"), crawlers.News)
 	})
 
-	engine.OnHTML(".hasTooltip", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+	engine.OnHTML(".PagedList-skipToNext > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
 		engine.Visit(element.Attr("href"), crawlers.Index)
 	})
